requirements.txt
psycopg2
pymongo
newspaper3k